I think I'm stable again.  Still in transition.
It took a bit of work to get the "get energy" routine working.  I think
I messed up one of the creep memory items and was trying to store a
container object instead of a container ID, and I was checking
"container.energy", which doesn't exist, instead of
"container.store[RESOURCE_ENERGY]".