There we go.  Properly fixed the energy refill spacing.
As long as I don't have the worker job system properly up and running, I
am making do.  My mod (%) operation to space out the workers between
various energy refill sources resulted in NaN if the worker number was 0
(0 vs number of refill targets => 0 is smaller => num targets % 0 =>
NaN).  Fixed it by force the denominator to 1 if it was 0.