Fixed a bug in the creep repair process.
The creep.memory.repairJobId wasn't null for some reason, and it wasn't
getting set to null, so it was looking for a structure that didn't
exist.  Forced it to null and made it reset, but now, disabling
structure repair (too expensive for every creep to look up all
structures on every tick, and no structures that decay ATM).  Also made
an adaptable function in spawn.queueWorkers to make more expensive
workers as more energy is available.