Converting creep job system from a queue system to a hash system
The job queues had a tendency to build up into large arrays when their
work wasn't getting done, which sometimes caused workers to stall as
they worked through the queue to get a valid job.  A hash-based system
takes advantage of JavaScript's nature as an object-based language with
hash keys to give each job its own object in a job collection.  Now each
job uses its unique object ID as a key to be unique in the job system.
Once a worker takes that job, the job is deleted and another job for the
same object can be submitted.  This still allows for jobs to build up,
but only amongst the creeps themselves.  There will only ever be one
copy of a job in waiting.